Modify the get_deadline function
Reduce the number of api calls by using the new file cache.json to store data locally. This file will store every deadline within a selected month, so it is only updated when user click on new month in the calendar. The cache.json is automatically deleted when the program end.